chore(toast): added toast config provider
